requirements.txt
Django==1.4.5
django-dajaxice
sqlalchemy==0.6.1
psycopg2==2.2.1
django-localeurl
django-staticfiles
jcc==3.00
